move effect controllers into the controllers/ directory
