+++ packages/flame/lib/src/effects2/controllers/effect_controller.dart
