+++ packages/flame/lib/src/effects2/controllers/simple_effect_controller.dart
