+++ packages/flame/lib/src/effects2/controllers/standard_effect_controller.dart
