+++ packages/flame/lib/src/effects2/remove_effect.dart
@@ -1,5 +1,5 @@
 import 'effect.dart';
-import 'simple_effect_controller.dart';
+import 'controllers/simple_effect_controller.dart';
 
 /// This simple effect, when attached to a component, will cause that component
 /// to be removed from the game tree after `delay` seconds.
