+++ packages/flame/lib/src/effects2/scale_effect.dart
@@ -1,6 +1,6 @@
 import 'package:vector_math/vector_math_64.dart';
 
-import 'effect_controller.dart';
+import 'controllers/effect_controller.dart';
 import 'transform2d_effect.dart';
 
 /// Scale a component.
