+++ packages/flame/test/effects2/effect_test.dart
@@ -1,7 +1,7 @@
 import 'package:flame/src/components/component.dart';
 import 'package:flame/src/effects2/effect.dart';
-import 'package:flame/src/effects2/effect_controller.dart';
-import 'package:flame/src/effects2/standard_effect_controller.dart';
+import 'package:flame/src/effects2/controllers/effect_controller.dart';
+import 'package:flame/src/effects2/controllers/standard_effect_controller.dart';
 import 'package:flame_test/flame_test.dart';
 import 'package:flutter_test/flutter_test.dart';
