+++ packages/flame/test/effects2/move_effect_test.dart
@@ -4,7 +4,7 @@ import 'dart:ui';
 import 'package:flame/components.dart';
 import 'package:flame/game.dart';
 import 'package:flame/src/effects2/move_effect.dart';
-import 'package:flame/src/effects2/simple_effect_controller.dart';
+import 'package:flame/src/effects2/controllers/simple_effect_controller.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
