+++ packages/flame/test/effects2/transform2d_effect_test.dart
@@ -1,6 +1,6 @@
 import 'package:flame/src/components/position_component.dart';
-import 'package:flame/src/effects2/effect_controller.dart';
-import 'package:flame/src/effects2/standard_effect_controller.dart';
+import 'package:flame/src/effects2/controllers/effect_controller.dart';
+import 'package:flame/src/effects2/controllers/standard_effect_controller.dart';
 import 'package:flame/src/effects2/transform2d_effect.dart';
 import 'package:flame_test/flame_test.dart';
 import 'package:flutter_test/flutter_test.dart';
